Code review feedback: - #1580
