+++ src/Microsoft.AspNet.Mvc.TagHelpers/TagHelperContextExtensions.cs
@@ -5,8 +5,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Framework.Logging;
-using Microsoft.AspNet.Mvc.TagHelpers.Internal;
 using Microsoft.AspNet.Mvc;
+using Microsoft.AspNet.Mvc.TagHelpers;
 
 namespace Microsoft.AspNet.Razor.Runtime.TagHelpers
 {
