+++ test/Microsoft.AspNet.Mvc.TagHelpers.Test/JavaScriptHelperTest.cs
@@ -6,39 +6,39 @@ using Xunit;
 
 namespace Microsoft.AspNet.Mvc.TagHelpers.Test
 {
-	public class JavaScriptHelperTest
-	{
-		[Theory]
-		[InlineData("Hello World", "Hello World")]
-		[InlineData("Hello & World", "Hello & World")]
-		[InlineData("Hello \r World", "Hello \\r World")]
-		[InlineData("Hello \n World", "Hello \\n World")]
-		[InlineData("Hello < World", "Hello \\u003c World")]
-		[InlineData("Hello > World", "Hello \\u003e World")]
-		[InlineData("Hello ' World", "Hello \\u0027 World")]
-		[InlineData("Hello \" World", "Hello \\u0022 World")]
-		[InlineData("Hello \\ World", "Hello \\\\ World")]
-		[InlineData("Hello \r\n <ah /> 'eep' \"hey\" World", "Hello \\r\\n \\u003cah /\\u003e \\u0027eep\\u0027 \\u0022hey\\u0022 World")]
-		public void JavaScriptEncode_EncodesCorrectly(string input, string expectedOutput)
-		{
-			// Act
-			var result = JavaScriptUtility.JavaScriptStringEncode(input);
-			
-			// Assert
-			Assert.Equal(expectedOutput, result);
-		}
-		
-		[Theory]
-		[InlineData("window.alert(\"[[[0]]]\")", "window.alert(\"{0}\")")]
-		[InlineData("var test = { a: 1 };", "var test = {{ a: 1 }};")]
-		[InlineData("var test = { a: 1, b: \"[[[0]]]\" };", "var test = {{ a: 1, b: \"{0}\" }};")]
-		public void PrepareFormatString_PreparesJavaScriptCorrectly(string input, string expectedOutput)
-		{
-			// Act
-			var result = JavaScriptUtility.PrepareFormatString(input);
-			
-			// Assert
-			Assert.Equal(expectedOutput, result);
-		}
-	}
+    public class JavaScriptHelperTest
+    {
+        [Theory]
+        [InlineData("Hello World", "Hello World")]
+        [InlineData("Hello & World", "Hello & World")]
+        [InlineData("Hello \r World", "Hello \\r World")]
+        [InlineData("Hello \n World", "Hello \\n World")]
+        [InlineData("Hello < World", "Hello \\u003c World")]
+        [InlineData("Hello > World", "Hello \\u003e World")]
+        [InlineData("Hello ' World", "Hello \\u0027 World")]
+        [InlineData("Hello \" World", "Hello \\u0022 World")]
+        [InlineData("Hello \\ World", "Hello \\\\ World")]
+        [InlineData("Hello \r\n <ah /> 'eep' \"hey\" World", "Hello \\r\\n \\u003cah /\\u003e \\u0027eep\\u0027 \\u0022hey\\u0022 World")]
+        public void JavaScriptEncode_EncodesCorrectly(string input, string expectedOutput)
+        {
+            // Act
+            var result = JavaScriptUtility.JavaScriptStringEncode(input);
+
+            // Assert
+            Assert.Equal(expectedOutput, result);
+        }
+
+        [Theory]
+        [InlineData("window.alert(\"[[[0]]]\")", "window.alert(\"{0}\")")]
+        [InlineData("var test = { a: 1 };", "var test = {{ a: 1 }};")]
+        [InlineData("var test = { a: 1, b: \"[[[0]]]\" };", "var test = {{ a: 1, b: \"{0}\" }};")]
+        public void PrepareFormatString_PreparesJavaScriptCorrectly(string input, string expectedOutput)
+        {
+            // Act
+            var result = JavaScriptUtility.PrepareFormatString(input);
+
+            // Assert
+            Assert.Equal(expectedOutput, result);
+        }
+    }
 }
